+++ Frameworks/PostgresKit/Source/FLXPostgresTypeNumberHandler.m
@@ -145,7 +145,7 @@ static FLXPostgresOid FLXPostgresTypeNumberTypes[] =
 	return nil;
 }
 
-- (id)objectFromResult:(const PGresult *)result atRow:(unsigned int)row column:(unsigned int)column
+- (id)objectFromResult:(const PGresult *)result atRow:(NSUInteger)row column:(NSUInteger)column
 {	
 	FLXPostgresOid type = PQftype(result, column);
 	NSUInteger length = PQgetlength(result, row, column);
